+++ src/ATGUI/atgui.cpp
@@ -10,91 +10,6 @@ ImColor Settings::UI::fontColor = ImColor(255, 255, 255, 255);
 
 #define IM_ARRAYSIZE(_ARR)  ((int)(sizeof(_ARR)/sizeof(*_ARR)))
 
-void UI::SetVisible(bool visible)
-{
-	UI::isVisible = visible;
-	cvar->FindVar("cl_mouseenable")->SetValue(!UI::isVisible);
-}
-
-void UI::SetupColors()
-{
-	ImGuiStyle& style = ImGui::GetStyle();
-
-	ImVec4 mainColorHovered	= ImVec4(Settings::UI::mainColor.Value.x + 0.1f, Settings::UI::mainColor.Value.y + 0.1f, Settings::UI::mainColor.Value.z + 0.1f, Settings::UI::mainColor.Value.w);
-	ImVec4 mainColorActive = ImVec4(Settings::UI::mainColor.Value.x + 0.2f, Settings::UI::mainColor.Value.y + 0.2f, Settings::UI::mainColor.Value.z + 0.2f, Settings::UI::mainColor.Value.w);
-	ImVec4 menubarColor = ImVec4(Settings::UI::bodyColor.Value.x, Settings::UI::bodyColor.Value.y, Settings::UI::bodyColor.Value.z, Settings::UI::bodyColor.Value.w - 0.8f);
-	ImVec4 frameBgColor = ImVec4(Settings::UI::bodyColor.Value.x, Settings::UI::bodyColor.Value.y, Settings::UI::bodyColor.Value.z, Settings::UI::bodyColor.Value.w + .1f);
-	ImVec4 tooltipBgColor = ImVec4(Settings::UI::bodyColor.Value.x, Settings::UI::bodyColor.Value.y, Settings::UI::bodyColor.Value.z, Settings::UI::bodyColor.Value.w + .05f);
-
-	style.Alpha = 1.0f;
-	style.WindowPadding = ImVec2(8, 8);
-	style.WindowMinSize = ImVec2(32, 32);
-	style.WindowRounding = 0.0f;
-	style.WindowTitleAlign = ImVec2(0.5f, 0.5f);
-	style.ChildWindowRounding = 0.0f;
-	style.FramePadding = ImVec2(4, 3);
-	style.FrameRounding = 0.0f;
-	style.ItemSpacing = ImVec2(8, 4);
-	style.ItemInnerSpacing = ImVec2(4, 4);
-	style.TouchExtraPadding = ImVec2(0, 0);
-	style.IndentSpacing = 21.0f;
-	style.ColumnsMinSpacing = 3.0f;
-	style.ScrollbarSize = 12.0f;
-	style.ScrollbarRounding = 0.0f;
-	style.GrabMinSize = 5.0f;
-	style.GrabRounding = 0.0f;
-	style.ButtonTextAlign = ImVec2(0.5f, 0.5f);
-	style.DisplayWindowPadding = ImVec2(22, 22);
-	style.DisplaySafeAreaPadding = ImVec2(4, 4);
-	style.AntiAliasedLines = true;
-	style.AntiAliasedShapes = true;
-	style.CurveTessellationTol = 1.25f;
-
-	style.Colors[ImGuiCol_Text] = Settings::UI::fontColor;
-	style.Colors[ImGuiCol_TextDisabled] = ImVec4(0.24f, 0.23f, 0.29f, 1.00f);
-	style.Colors[ImGuiCol_WindowBg] = Settings::UI::bodyColor;
-	style.Colors[ImGuiCol_ChildWindowBg] = ImVec4(.0f, .0f, .0f, .0f);
-	style.Colors[ImGuiCol_PopupBg] = tooltipBgColor;
-	style.Colors[ImGuiCol_Border] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_BorderShadow] = ImVec4(0.92f, 0.91f, 0.88f, 0.00f);
-	style.Colors[ImGuiCol_FrameBg] = frameBgColor;
-	style.Colors[ImGuiCol_FrameBgHovered] = mainColorHovered;
-	style.Colors[ImGuiCol_FrameBgActive] = mainColorActive;
-	style.Colors[ImGuiCol_TitleBg] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_TitleBgCollapsed] = ImVec4(1.00f, 0.98f, 0.95f, 0.75f);
-	style.Colors[ImGuiCol_TitleBgActive] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_MenuBarBg] = menubarColor;
-	style.Colors[ImGuiCol_ScrollbarBg] = ImVec4(frameBgColor.x + .05f, frameBgColor.y + .05f, frameBgColor.z + .05f, frameBgColor.w);
-	style.Colors[ImGuiCol_ScrollbarGrab] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_ScrollbarGrabHovered] = mainColorHovered;
-	style.Colors[ImGuiCol_ScrollbarGrabActive] = mainColorActive;
-	style.Colors[ImGuiCol_ComboBg] = frameBgColor;
-	style.Colors[ImGuiCol_CheckMark]  = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_SliderGrab] = mainColorHovered;
-	style.Colors[ImGuiCol_SliderGrabActive] = mainColorActive;
-	style.Colors[ImGuiCol_Button] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_ButtonHovered] = mainColorHovered;
-	style.Colors[ImGuiCol_ButtonActive] = mainColorActive;
-	style.Colors[ImGuiCol_Header] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_HeaderHovered] = mainColorHovered;
-	style.Colors[ImGuiCol_HeaderActive] = mainColorActive;
-	style.Colors[ImGuiCol_Column] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_ColumnHovered] = mainColorHovered;
-	style.Colors[ImGuiCol_ColumnActive] = mainColorActive;
-	style.Colors[ImGuiCol_ResizeGrip] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_ResizeGripHovered] = mainColorHovered;
-	style.Colors[ImGuiCol_ResizeGripActive] = mainColorActive;
-	style.Colors[ImGuiCol_CloseButton] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_CloseButtonHovered] = mainColorHovered;
-	style.Colors[ImGuiCol_CloseButtonActive] = mainColorActive;
-	style.Colors[ImGuiCol_PlotLines] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_PlotLinesHovered] = mainColorHovered;
-	style.Colors[ImGuiCol_PlotHistogram] = Settings::UI::mainColor;
-	style.Colors[ImGuiCol_PlotHistogramHovered] =  mainColorHovered;
-	style.Colors[ImGuiCol_TextSelectedBg] = mainColorHovered;
-	style.Colors[ImGuiCol_ModalWindowDarkening] = ImVec4(0.0f, 0.0f, 0.0f, 0.75f);
-}
-
 void SetupMainMenuBar()
 {
 	if (ImGui::BeginMainMenuBar())
@@ -135,6 +50,12 @@ void UI::SwapWindow()
 		return;
 
 	Draw::ImDrawText(ImVec2(4.f, 4.f), ImColor(255, 255, 255, 255), "AimTux", NULL, 0.0f, NULL, ImFontFlags_Shadow);
+}
+
+void UI::SetVisible(bool visible)
+{
+	UI::isVisible = visible;
+	cvar->FindVar("cl_mouseenable")->SetValue(!UI::isVisible);
 }
 
 void UI::SetupWindows()
